requirements.txt
pandas
numpy
#pyspark
paramiko
fastapi
uvicorn[standard]
sqlalchemy
pydantic
loguru
python-dateutil
slowapi
pytest
requests
httpx